urls와 users 폴더를 만들고 urls기능에 맞는 view등을 넣어줌
