+++ shortener/urls/urls.py
@@ -1,0 +1,9 @@
+from shortener.urls.views import url_change, url_list, url_create
+from django.urls import path
+
+
+urlpatterns = [
+    path("", url_list, name="url_list"),
+    path("create", url_create, name="url_create"),
+    path("<str:action>/<int:url_id>", url_change, name="url_change"),
+]
