+++ shortener/urls/views.py
@@ -1,0 +1,78 @@
+from django.shortcuts import render, redirect
+from shortener.forms import UrlCreateForm
+from shortener.models import ShortenedUrls
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+
+from shortener.utils import url_count_changer
+
+
+@login_required
+def url_list(request):
+    get_list = ShortenedUrls.objects.order_by("-created_at").all()
+    context = {"list": get_list}
+    return render(request, "url_list.html", context)
+
+
+@login_required
+def url_create(request):
+    msg = None
+    if request.method == "POST":
+        form = UrlCreateForm(request.POST)
+
+        if form.is_valid():
+            msg = f"{form.cleaned_data.get('nick_name')} 생성 완료!"
+            messages.add_message(request, messages.INFO, msg)
+            form.save(request)
+            return redirect("url_list")
+
+        else:
+            form = UrlCreateForm()
+
+    else:
+        form = UrlCreateForm()
+
+    context = {"form": form}
+    return render(request, "url_create.html", context)
+
+
+@login_required
+def url_change(request, action, url_id):
+    if request.method == "POST":
+        url_data = ShortenedUrls.objects.filter(id=url_id)
+
+        if url_data.exists():
+            if url_data.first().creator_id != request.user.id:
+                msg = "자신이 소유하지 않은 URL 입니다."
+
+            else:
+                if action == "delete":
+                    msg = f"{url_data.first().nick_name} 삭제 완료!"
+
+                    try:
+                        url_data.delete()
+                    except Exception as e:
+                        print(e)
+                    else:
+                        url_count_changer(request, False)
+
+                    messages.add_message(request, messages.INFO, msg)
+
+                elif action == "update":
+                    msg = f"{url_data.first().nick_name} 수정 완료!"
+                    form = UrlCreateForm(request.POST)
+                    form.update_form(request, url_id)
+
+                    messages.add_message(request, messages.INFO, msg)
+
+        else:
+            msg = "해당 URL 정보를 찾을 수 없습니다."
+
+    elif request.method == "GET" and action == "update":
+        url_data = ShortenedUrls.objects.filter(pk=url_id).first()
+        form = UrlCreateForm(instance=url_data)
+
+        context = {"form": form, "is_update": True}
+        return render(request, "url_create.html", context)
+
+    return redirect("url_list")
