+++ shortener/views.py
@@ -1,11 +1,9 @@
 from django.http import JsonResponse
 from django.shortcuts import render, redirect
-from shortener.forms import LoginForm, RegisterForm, UrlCreateForm
-from shortener.models import Users, ShortenedUrls
+from shortener.forms import LoginForm, RegisterForm
+from shortener.models import Users
 from django.views.decorators.csrf import csrf_exempt
-from django.contrib import messages
 from django.contrib.auth import login, authenticate, logout
-from django.contrib.auth.forms import AuthenticationForm, PasswordResetForm, PasswordChangeForm
 from django.core.paginator import Paginator
 from django.contrib.auth.decorators import login_required
 
@@ -67,15 +65,15 @@ def login_view(request):
             msg = "올바른 유저ID와 패스워드를 입력하세요."
 
             try:
-                user = Users.objects.get(email=email)
+                user = Users.objects.get(user__email=email)
 
             except Users.DoesNotExist:
                 pass
 
             else:
-                if user.check_password(raw_password):
+                if user.user.check_password(raw_password):
                     msg = None
-                    login(request, user)
+                    login(request, user.user)
                     is_ok = True
                     request.session["remember_me"] = remember_me
 
@@ -103,67 +101,3 @@ def list_view(request):
 
     context = {"users": users}
     return render(request, "boards.html", context)
-
-
-@login_required
-def url_list(request):
-    get_list = ShortenedUrls.objects.order_by("-created_at").all()
-    context = {"list": get_list}
-    return render(request, "url_list.html", context)
-
-
-@login_required
-def url_create(request):
-    msg = None
-    if request.method == "POST":
-        form = UrlCreateForm(request.POST)
-
-        if form.is_valid():
-            msg = f"{form.cleaned_data.get('nick_name')} 생성 완료!"
-            messages.add_message(request, messages.INFO, msg)
-            form.save(request)
-            return redirect("url_list")
-
-        else:
-            form = UrlCreateForm()
-
-    else:
-        form = UrlCreateForm()
-
-    context = {"form": form}
-    return render(request, "url_create.html", context)
-
-
-@login_required
-def url_change(request, action, url_id):
-    if request.method == "POST":
-        url_data = ShortenedUrls.objects.filter(id=url_id)
-
-        if url_data.exists():
-            if url_data.first().created_by_id != request.user.id:
-                msg = "자신이 소유하지 않은 URL 입니다."
-
-            else:
-                if action == "delete":
-                    msg = f"{url_data.first().nick_name} 삭제 완료!"
-                    url_data.delete()
-                    messages.add_message(request, messages.INFO, msg)
-
-                elif action == "update":
-                    msg = f"{url_data.first().nick_name} 수정 완료!"
-                    form = UrlCreateForm(request.POST)
-                    form.update_form(request, url_id)
-
-                    messages.add_message(request, messages.INFO, msg)
-
-        else:
-            msg = "해당 URL 정보를 찾을 수 없습니다."
-
-    elif request.method == "GET" and action == "update":
-        url_data = ShortenedUrls.objects.filter(pk=url_id).first()
-        form = UrlCreateForm(instance=url_data)
-
-        context = {"form": form, "is_update": True}
-        return render(request, "url_create.html", context)
-
-    return redirect("url_list")
