+++ shrinkers/urls.py
@@ -20,17 +20,16 @@ from shrinkers.settings import DEBUG
 #     import debug_toolbar
 from django.contrib import admin
 from django.urls import path
+from django.conf.urls import include
+
 from shortener.views import (
-    index, get_user,
+    index,
+    get_user,
     list_view,
     register,
     login_view,
     logout_view,
-    url_list,
-    url_create,
-    url_change
 )
-from django.conf.urls import include
 
 
 urlpatterns = [
@@ -40,10 +39,8 @@ urlpatterns = [
     path("login", login_view, name="login"),
     path("logout", logout_view, name="logout"),
     path("list", list_view, name="list_view"),
-    path("urls", url_list, name="url_list"),
-    path("urls/create", url_create, name="url_create"),
-    path("urls/<str:action>/<int:url_id>", url_change, name="url_change"),
     path("get_user/<int:user_id>", get_user),
+    path("urls/", include("shortener.urls.urls")),
 ]
 
 # if DEBUG:
