+++ templates/url_list.html
@@ -33,7 +33,7 @@
               <th scope="row">{{ l.nick_name }}</th>
               <td>{{ l.target_url }}</td>
               <td>{{ l.shortened_url }}</td>
-              <td>{{ l.created_via }}</td>
+              <td>{{ l.create_via }}</td>
               <td>{{ l.created_at }}</td>
               <td>
                 <form action="/urls/delete/{{ l.id }}" method="POST">
